+++ src/queries/index.js
@@ -1,5 +1,6 @@
 export { default as getUserTenantsPermissions } from './getUserTenantsPermissions';
 export { default as useChunkedCQLFetch } from './useChunkedCQLFetch';
+export { default as useChunkedIdTransformFetch } from './useChunkedIdTransformFetch';
 export { default as useConfigurations } from './useConfigurations';
 export { default as useOkapiEnv } from './useOkapiEnv';
 export { default as useSettings } from './useSettings';
